% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Basic scRNA-seq QC metrics from an ovarian tumor}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
    \item{detected}{number of unique genes detected in that cell's transcripts}
    \item{subsets_mito_percent}{percent of reads mapping to mitochondrial genes}
    ...
}
}
\source{
GEO Accession GSM4816047
}
\usage{
metrics
}
\description{
Basic QC metrics from a high-grade serous ovarian cancer (HGSOC) sample.
The information included is the minimum needed to generate a miQC model and
make plots. Count data for the full tumor is available through GEO,
accession number GSM4816047.
}
\keyword{datasets}
