% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureModel.R
\name{mixtureModel}
\alias{mixtureModel}
\title{mixtureModel}
\usage{
mixtureModel(sce, model_type = "linear")
}
\arguments{
\item{sce}{(SingleCellExperiment) Input data object.}

\item{model_type}{(character) What type of model to generate. A linear
mixture model ("linear") is recommended, but currently b-spline ("spline")
and two-degree polynomial ("polynomial") are also supported
Default = "linear".}
}
\value{
Returns a flexmix object with mixture model parameters, which is used
  to calculate posterior probability for each cell being compromised and make
  final filtering decisions.
}
\description{
Function to fit a two-distribution mixture model on a SingleCellExperiment
object.
}
\examples{
library(scRNAseq)
sce <- ZeiselBrainData()
mt_genes <- grepl("^mt-",  rownames(sce))
feature_ctrls <- list(mito = rownames(sce)[mt_genes])
sce <- addPerCellQC(sce, subsets = feature_ctrls, BPPARAM = BiocParallel::MulticoreParam())
model <- mixtureModel(sce)
}
