% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFiltering.R
\name{plotFiltering}
\alias{plotFiltering}
\title{plotFiltering}
\usage{
plotFiltering(
  sce,
  model = NULL,
  posterior_cutoff = 0.75,
  palette = c("#999999", "#E69F00")
)
}
\arguments{
\item{sce}{(SingleCellExperiment) Input data object.}

\item{model}{(flexmix) Output of mixtureModel function, which should be
explicitly called first to ensure stability of model parameters.
Default = NULL.}

\item{posterior_cutoff}{(numeric) The posterior probability of a cell being
part of the compromised distribution, a number between 0 and 1. Any cells
below the appointed cutoff will be marked to keep.
Default = 0.75}

\item{palette}{(character) Color palette. A vector of length two containing
custom colors.
Default = c("#999999", "#E69F00").}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, customized formatting, etc).
}
\description{
Function to plot which cells will be kept and removed given their posterior
probability of belonging to the compromised distribution.
}
\examples{
library(scRNAseq)
sce <- ZeiselBrainData()
mt_genes <- grepl("^mt-",  rownames(sce))
feature_ctrls <- list(mito = rownames(sce)[mt_genes])
sce <- addPerCellQC(sce, subsets = feature_ctrls, BPPARAM = BiocParallel::MulticoreParam())
model <- mixtureModel(sce)
sce <- plotFiltering(sce, model)
}
